requirements.txt
aiohttp==3.9.1
aioitertools==0.11.0
aiobotocore==2.11.1
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==4.1.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.8.1
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-postgres==5.8.0
apache-airflow-providers-sqlite==3.5.0
apache-airflow-providers-telegram==4.2.0
apispec==6.3.0
argcomplete==3.1.6
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
blinker==1.7.0
boto3==1.34.27
botocore==1.34.27
build==1.0.3
CacheControl==0.13.1
cachelib==0.9.0
catboost==1.2.2
category-encoders==2.6.3
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
crashtest==0.4.1
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
Deprecated==1.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dvc==3.42.0
dvc-data==3.8.0
dvc-http==2.32.0
dvc-objects==3.0.6
dvc-render==1.0.1
dvc-s3==3.0.1
dvc-studio-client==0.18.0
dvc-task==0.3.0
entrypoints==0.4
dill==0.3.7
distlib==0.3.7
dnspython==2.4.2
docutils==0.20.1
dulwich==0.21.6
email-validator==1.3.1
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.0
google-re2==1.1
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keyring==24.3.0
lazy-object-proxy==1.9.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prison==0.2.1
protobuf==4.25.1
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
python-telegram-bot==20.7
python-dotenv==1.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
rapidfuzz==3.5.2
referencing==0.31.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.13.2
setproctitle==1.3.3
scikit-learn==1.4.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.3
trove-classifiers==2023.11.29
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==2.0.7
virtualenv==20.25.0
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.0.1
xattr==0.10.1
yarl==1.9.3
zipp==3.17.0
